+++ shell/cmd_uptime.c
@@ -5,7 +5,7 @@ static int cmd_uptime(const char *args[])
 {
 	extern uint32_t uptime_sec;
 
-	pp_printf("%u\n", uptime_sec);
+	pp_printf("%u\n", (unsigned int) uptime_sec);
 	return 0;
 }
